# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pickup Sports App API
  version: 1.0.0
  description: |
    REST API for Pickup Sports App (Java 17, Spring Boot). Includes authentication, game management, search, notifications, profiles, venues, Nepal market features, media presign, and push subscriptions. JWT Bearer auth used for protected endpoints. Some features may be gated by feature flags or env properties.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/games/nearby:
    get:
      tags: [Games]
      summary: Find nearby games
      parameters:
        - in: query
          name: latitude
          required: true
          schema: { type: number, format: double }
        - in: query
          name: longitude
          required: true
          schema: { type: number, format: double }
        - in: query
          name: radiusKm
          schema: { type: number, format: double, default: 5.0 }
        - in: query
          name: sport
          schema: { type: string }
        - in: query
          name: skillLevel
          schema: { type: string }
      responses:
        '200':
          description: Games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameSummaryDTO'
...
  /venues/nearby:
    get:
      tags: [Venues]
      summary: Find nearby venues
      parameters:
        - in: query
          name: latitude
          required: true
          schema: { type: number }
        - in: query
          name: longitude
          required: true
          schema: { type: number }
        - in: query
          name: radiusKm
          required: true
          schema: { type: number }
      responses:
        '200':
          description: Venues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VenueResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GameSummaryDTO:
      type: object
      properties:
        id: { type: integer, format: int64 }
        sport: { type: string }
        location: { type: string }
        time: { type: string, format: date-time }
        skillLevel: { type: string }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        creatorName: { type: string }
        currentPlayers: { type: integer }
        maxPlayers: { type: integer }
        status: { type: string }
...
    VenueResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        description: { type: string }
        address: { type: string }
        city: { type: string }
        state: { type: string }
        country: { type: string }
        postalCode: { type: string }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        venueType: { type: string }
        maxCapacity: { type: integer }
        minCapacity: { type: integer }
        status: { type: string }
        basePricePerHour: { type: number, format: double }
        basePricePerPlayer: { type: number, format: double }
        isVerified: { type: boolean }
        ownerId: { type: integer, format: int64 }
        contactPhone: { type: string }
        contactEmail: { type: string }
        websiteUrl: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        supportedSports:
          type: array
          items:
            type: object
        amenities:
          type: array
          items:
            type: object
        businessHours:
          type: array
          items:
            type: object
        images:
          type: array
          items:
            type: object
...